Actualizacion maquinas y alimento

+++ ProyectoGimnacio/ProyectoGimnacio/Clases/ClsNutricion.cs
@@ -10,9 +10,9 @@ namespace ProyectoGimnacio.Clases
         public int AlimentoID { get; set; }
         public string Nombre { get; set; }
         public int Calorias { get; set; }
-        public int Proteinas { get; set; }
+        public decimal Proteinas { get; set; }
         public decimal Grasas { get; set; }
-        public int Carbohidratos { get; set; }
+        public decimal Carbohidratos { get; set; }
         public int Peso { get; set; }
     }
 
@@ -21,9 +21,9 @@ namespace ProyectoGimnacio.Clases
         public int ComidaID { get; set; }
         public string Nombre { get; set; }
         public int Calorias { get; set; }
-        public int Proteinas { get; set; }
+        public decimal Proteinas { get; set; }
         public decimal Grasas { get; set; }
-        public int Carbohidratos { get; set; }
+        public decimal Carbohidratos { get; set; }
     }
 
     public class Menu
